refactor(ui): extract NicebarMixin for the inline notification strip
Move show_nicebar / hide_nicebar plus the inline stylesheet into
l5r/ui/nicebar.py. L5RMain.__init__ still initialises self.nicebar = None
and build_ui still creates self.mvbox — the mixin reads both.

main.py: 2406 -> 2361 lines.
