+++ l5r/main.py
@@ -54,11 +54,12 @@ from l5r.ui.helpers import (
     new_small_plus_bt,
     new_vert_line,
 )
+from l5r.ui.nicebar import NicebarMixin
 from l5r.util import log
 from l5r.util.settings import L5RCMSettings
 
 
-class L5RMain(AdviseMixin, L5RCMCore):
+class L5RMain(AdviseMixin, NicebarMixin, L5RCMCore):
 
     default_size = QtCore.QSize(820, 720)
     default_point_size = 8.25
@@ -1485,52 +1486,6 @@ class L5RMain(AdviseMixin, L5RCMCore):
         self.bt_edit_school.clicked.connect(self.sink4.on_edit_first_school)
 
         self.bt_set_exp_points.clicked.connect(self.sink1.on_set_exp_limit)
-
-    def show_nicebar(self, wdgs):
-        self.nicebar = QtWidgets.QFrame(self)
-        self.nicebar.setStyleSheet("""
-        QWidget { background: beige;}
-        QPushButton {
-            color: #333;
-            border: 2px solid rgb(200,200,200);
-            border-radius: 7px;
-            padding: 5px;
-            background: qradialgradient(cx: 0.3, cy: -0.4,
-            fx: 0.3, fy: -0.4, radius: 1.35, stop: 0 #fff,
-            stop: 1 rgb(255,170,0));
-            min-width: 80px;
-            }
-
-        QPushButton:hover {
-            background: qradialgradient(cx: 0.3, cy: -0.4,
-            fx: 0.3, fy: -0.4, radius: 1.35, stop: 0 #fff,
-            stop: 1 rgb(255,100,30));
-        }
-
-        QPushButton:pressed {
-            background: qradialgradient(cx: 0.4, cy: -0.1,
-            fx: 0.4, fy: -0.1, radius: 1.35, stop: 0 #fff,
-            stop: 1 rgb(255,200,50));
-        }
-        """)
-        self.nicebar.setMinimumSize(0, 32)
-
-        # nicebar layout
-        hbox = QtWidgets.QHBoxLayout(self.nicebar)
-        hbox.setContentsMargins(9, 1, 9, 1)
-
-        for w in wdgs:
-            hbox.addWidget(w)
-
-        self.mvbox.insertWidget(1, self.nicebar)
-        self.nicebar.setVisible(True)
-
-    def hide_nicebar(self):
-        if not self.nicebar:
-            return
-        self.nicebar.setVisible(False)
-        del self.nicebar
-        self.nicebar = None
 
     def on_trait_increase(self, tag):
         """raised when user click on the small '+' button near traits"""
